requirements.txt
scikit-learn
numpy
PyPDF2
python-docx
streamlit
openpyxl
reportlab
